feat: 拆分 UI 模組並更新聚合入口
- 將 ui.js 拆分為 shared/ui/{toolbar,status,classes,logIcons}.js
- 根目錄 ui.js 轉為 façade
- 更新 shared/ui/index.js 和 shared/index.js 聚合入口
- 修正 env.example.js façade 加入 runtime 匯出以維持測試相容
