+++ scripts/config/shared/index.js
@@ -2,3 +2,5 @@ export * from './status/index.js';
 export * from './notion/index.js';
 export * from './highlight/index.js';
 export * from './ui/index.js';
+export * from './core/index.js';
+export * from './storage/index.js';
