+++ scripts/config/shared/ui/classes.js
@@ -1,0 +1,12 @@
+/**
+ * 通用 UI 組件 CSS 類名
+ */
+
+export const COMMON_CSS_CLASSES = {
+  SUCCESS_BOX: 'success-box',
+  ERROR_BOX: 'error-box',
+  WARNING_BOX: 'warning-box',
+  RESULT_URL: 'result-url',
+  RESULT_ITEM: 'migration-result-item',
+  COUNT_BADGE: 'count-badge',
+};
