+++ scripts/config/shared/ui/index.js
@@ -1,1 +1,9 @@
+/**
+ * UI 配置聚合入口
+ */
+
 export * from './icons.js';
+export * from './toolbar.js';
+export * from './status.js';
+export * from './classes.js';
+export * from './logIcons.js';
