+++ scripts/config/shared/ui/logIcons.js
@@ -1,0 +1,12 @@
+/**
+ * 統一日誌圖標定義 (Centralized Emoji Config)
+ * 用於 Logger.success/start/ready 等語義化方法
+ */
+
+export const LOG_ICONS = {
+  SUCCESS: '✅',
+  ERROR: '❌',
+  WARN: '⚠️',
+  START: '🚀',
+  READY: '📦',
+};
