+++ scripts/config/shared/ui/status.js
@@ -1,0 +1,10 @@
+/**
+ * UI 提示狀態類型
+ */
+
+export const UI_STATUS_TYPES = {
+  SUCCESS: 'success',
+  ERROR: 'error',
+  INFO: 'info',
+  WARNING: 'warning',
+};
