+++ scripts/config/shared/ui/toolbar.js
@@ -1,0 +1,17 @@
+/**
+ * Toolbar DOM 選擇器
+ */
+
+export const TOOLBAR_SELECTORS = {
+  CONTAINER: '#notion-highlighter-v2',
+  MINI_ICON: '#notion-highlighter-mini-icon',
+  TOGGLE_HIGHLIGHT: '#toggle-highlight-v2',
+  MINIMIZE: '#minimize-highlight-v2',
+  CLOSE: '#close-highlight-v2',
+  COLOR_PICKER: '#color-picker-v2',
+  SYNC_TO_NOTION: '#sync-to-notion-v2',
+  MANAGE_HIGHLIGHTS: '#manage-highlights-v2',
+  SAVE_PAGE: '#save-page-v2',
+  STATUS_CONTAINER: '#highlight-status-v2',
+  COUNT_DISPLAY: '#sync-count-badge-v2',
+};
